+++ Setup-Your-Mac-via-Dialog.bash
@@ -64,7 +64,7 @@
 #   Version 1.12.8, 13-Sep-2023, Dan K. Snelson (@dan-snelson)
 #   - Added a check for FileVault being enabled during Setup Assistant (for macOS 14 Sonoma) ([Pull Request No. 96](https://github.com/dan-snelson/Setup-Your-Mac/pull/96); thanks, Obi-@drtaru!)
 #
-#   Version 1.12.9, 14-Sep-2023, Dan K. Snelson (@dan-snelson)
+#   Version 1.12.9, 15-Sep-2023, Dan K. Snelson (@dan-snelson)
 #   - Added `-L` to `curl` command when caching banner images (thanks for the suggestion, @bartreardon!)
 #   - Added `swiftDialogMinimumRequiredVersion` variable to more easily track the minimum build. ([Pull Request No. 98](https://github.com/dan-snelson/Setup-Your-Mac/pull/98); thanks, @GadgetGeekNI!)
 #   - Hide unused Support variables ([Pull Request No. 99](https://github.com/dan-snelson/Setup-Your-Mac/pull/99); thanks again, @GadgetGeekNI!)
@@ -84,7 +84,7 @@
 # Script Version and Jamf Pro Script Parameters
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
 
-scriptVersion="1.12.9-rc3"
+scriptVersion="1.12.9"
 export PATH=/usr/bin:/bin:/usr/sbin:/sbin
 scriptLog="${4:-"/var/log/org.churchofjesuschrist.log"}"                        # Parameter 4: Script Log Location [ /var/log/org.churchofjesuschrist.log ] (i.e., Your organization's default location for client-side logs)
 debugMode="${5:-"verbose"}"                                                     # Parameter 5: Debug Mode [ verbose (default) | true | false ]