+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # CHANGELOG
 
 ## 1.12.9
-### 14-Sep-2023
+### 15-Sep-2023
 - Added `-L` to `curl` command when caching banner images (thanks for the suggestion, @bartreardon!)
 - Added `swiftDialogMinimumRequiredVersion` variable to more easily track the minimum build. ([Pull Request No. 98](https://github.com/dan-snelson/Setup-Your-Mac/pull/98); thanks, @GadgetGeekNI!)
 - Hide unused Support variables ([Pull Request No. 99](https://github.com/dan-snelson/Setup-Your-Mac/pull/99); thanks again, @GadgetGeekNI!)